Commit before changing to an EXE.
Using a new AppDomain seemed to be working, but now we have an error from
the task saying it cannot load SimpleInjector due to a serialization failure,
something to do with the Logger in the Builder and trying to cross an
AppDomain.

There are obviously too many problems which might bite me down the road if
I continue to follow this approach, so I am going to convert Lithogen to
use an EXE instead, and have a task shim call that EXE with the correct
command line parameters.